refactor: Enable `OpoInterpreter` reuse in other projects
This small change moves towards allowing the OpoLua core functionality to be reused in other projects by marking some functions and classes as public and moving the `CGImage` extensions into the 'swift' directory so this can be use standalone (alongside the 'src' directory which contains the corresponding Lua source).

It's unclear what the long-term approach to allowing code reuse across projects will be, but this allows for initial experimentation.
