+++ swift/OpoIoHandler_CoreGraphics.swift
@@ -19,7 +19,6 @@
 // SOFTWARE.
 
 import CoreGraphics
-import UIKit
 
 extension Graphics.Color {
     func cgColor() -> CGColor {
@@ -42,41 +41,5 @@ extension Graphics.Point {
 extension Graphics.Rect {
     func cgRect() -> CGRect {
         return CGRect(x: self.origin.x, y: self.origin.y, width: self.width, height: self.height)
-    }
-}
-
-extension Graphics.FontInfo {
-    func toUiFont() -> UIFont? {
-        let sz = CGFloat(self.size)
-        let uiFontName: String
-        var traits: UIFontDescriptor.SymbolicTraits = []
-        if self.flags.contains(.bold) || self.flags.contains(.boldHint) {
-            traits.insert(.traitBold)
-        }
-        switch self.face {
-        case .arial:
-            uiFontName = "Arial"
-        case .times:
-            uiFontName = "Times"
-        case .courier:
-            uiFontName = "Courier"
-        case .tiny:
-            uiFontName = "Courier" // Who knows...
-        case .squashed:
-            uiFontName = "Helvetica Neue"
-            traits.insert(.traitCondensed)
-        case .digit, .eiksym:
-            return nil
-        }
-
-        var desc = UIFontDescriptor(name: uiFontName, size: sz)
-        if let newDesc = desc.withSymbolicTraits(traits) {
-            desc = newDesc
-        }
-        return UIFont(descriptor: desc, size: sz)
-    }
-
-    func toBitmapFont() -> BitmapFontInfo? {
-        return BitmapFontInfo(uid: self.uid)
     }
 }
